// src/types/theme.ts
export const themeStylesSchema = z.object({
  light: themeStylePropsSchema,
  dark: themeStylePropsSchema,
});

export type ThemeStyleProps = z.infer<typeof themeStylePropsSchema>;
export type ThemeStyles = z.infer<typeof themeStylesSchema>;
// ...
export type Theme = {
  id: string;
  userId: string;
  name: string;
  styles: ThemeStyles;
  createdAt: Date;
  updatedAt: Date;
};